A seperate rpc file is now used for the voting contract calls. This file should hold all the gridcoin contract rpc's in the future.
Remove the direct rpc call from votingdialog. Just like the rest of qt,
besides the rpc console, it should use native functions instead of rpc.
